Fixed additional unsupported inline asm calls.
This commit is an extension to commit 5deb901177aa2ab056188b76bafae083935d7a5b. This fixes
some additional inline asm calls. These changes ensure compatibility with Clang and GCC.

+++ c/ldv-linux-3.4-simple/32_1_cilled_true-unreach-call_ok_nondet_linux-3.4-32_1-drivers--mtd--devices--doc2001plus.ko-ldv_main0_sequence_infinite_withcheck_stateful.cil.out.c
@@ -2581,7 +2581,7 @@ long ldv__builtin_expect(long val , long res ) ;
 __inline static int ffs(int x )  __attribute__((__no_instrument_function__)) ;
 #line 397 "/home/zakharov/launch/inst/current/envs/linux-3.4/linux-3.4/arch/x86/include/asm/bitops.h"
 __inline static int ffs(int x ) 
-{ int r ;
+{ long r ;
   long tmp ;
 
   {
